Separate the Realization class from Tuple.h
